moved more functions out of met_util
